Refactor db seeding to use prisma
